fix: don't create extra dir at output path
previously if you tried to index a package at dir `foo` to path `bar` it
would write to `bar/foo` instead of just `bar` which is not really what
you usually want. For example in zola you'd want the content folder to
be the output and having an extra dir in there breaks the website
basically.